+++ HR_Managment/HR_Managment.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HR_Managment.Application.Features.LeaveTypes.Requests.Commands;
+﻿using HR_Managment.Application.Exceptions;
+using HR_Managment.Application.Features.LeaveTypes.Requests.Commands;
 using HR_Managment.Application.Persistence.Contracts;
+using HR_Managment.Domain;
 using MapsterMapper;
 using MediatR;
 
@@ -18,6 +20,10 @@ public class DeleteLeaveTypeCommandHandler
     }
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
+        var leaveType = await _leaveTypeRepository.Get(request.Id);
+        if (leaveType == null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
+
         await _leaveTypeRepository.Delete(request.Id);
         return Unit.Value;
     }
